refactor(core): split CommandRouter into partial files (938 -> 668 lines)
Fix #224 (focused extraction).

CommandRouter.cs grew to 938 lines because it carried 14 nearly-identical
Queue*ForLockstep helper methods (~280 lines) right next to the public
Issue* API and the routing decisions. The helpers are pure boilerplate —
each one fetches network IDs, falls back to local execution if missing,
and wraps a LockstepCommand for the service locator.

Marked CommandRouter as `partial class` and moved the entire LOCKSTEP
QUEUE METHODS section to a sibling partial file:

  CommandRouter.LockstepQueue.cs (307 lines)
    QueueMoveForLockstep, QueueAttackForLockstep, QueueAttackMoveForLockstep,
    QueueStopForLockstep, QueueHoldPositionForLockstep, QueueBuildForLockstep,
    QueueGatherForLockstep, QueueHealForLockstep, QueueRallyPointForLockstep,
    QueueRepairForLockstep, QueuePatrolForLockstep, QueueConvertForLockstep,
    QueueTrainForLockstep, QueueAbilityForLockstep

The main CommandRouter.cs is now 668 lines and owns:
  - The public Issue* API (the entry point for player/AI/system commands)
  - The CommandSource enum + ShouldQueueForLockstep routing decision
  - GetNetworkId helper (consumed by both partials)
  - The direct-execution fallbacks (SetRallyPointDirect, TrainCommandDirect,
    IssueAbilityDirect, PlaceBuildingDirect)

Both partials live in the same namespace and class, so all access stays
private static and no public surface changed. GetNetworkId stays in the
main file as the shared helper.

Closes #224 (focused extraction; remaining 500+ line files documented as
incremental follow-up — many are tracked by separate dedicated tasks)
